Improve WeatherDetails with code cleanup and enhanced API calls
Removed unnecessary imports in router/index.ts largely improving the overall performance of the application. Refactored the WeatherDetails component to no longer use hard-coded city detail and instead use dynamic data fetched from the API. This involved changing the parameter used for city details and upgrading the fetchWeatherHistory method to first get the latitude and longitude of the city and then get the forecast for the next 5 days using that data.